+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+### 3.1.44
+
+- Games added:
+    - Atomicrops
+    - Erenshor
+    - Sunkenland
+    - Wizard with a Gun (Thanks to @RandomWolf)
+
+### 3.1.43
+All changes were by the Thunderstore team for this release
+
+- Games added:
+  - Wizard of Legend
+  - Will You Snail?
+  - Garfield Kart - Furious Racing
+  - Techtonica
+  - Thronefall
+  - We Love Katamari REROLL + Royal Reverie
+  - Bomb Rush Cyberfunk
+  - Touhou: Lost Branch of Legend
+
 ### 3.1.42
 - Games using MelonLoader should now correctly generate assemblies on first launch (EG: BONEWORKS / BONELABS)
 - Data directory can (finally) be re-set to an older directory without having to clear settings